requirements.txt
groq==0.22.0
python-dotenv>=1.0.0
pydantic==2.11.3
typing-extensions==4.13.2
tenacity==9.1.2
openai-agents==0.0.11
requests==2.31.0
tavily-python==0.5.4
